refactor(editpage): change where context lives
