+++ src/layouts/EditPage/MarkdownEditPage.tsx
@@ -10,42 +10,26 @@ import {
   ModalFooter,
   ModalHeader,
   ModalOverlay,
+  Alert,
 } from "@chakra-ui/react"
 import { Button } from "@opengovsg/design-system-react"
-import BubbleMenu from "@tiptap/extension-bubble-menu"
-import CharacterCount from "@tiptap/extension-character-count"
-import Highlight from "@tiptap/extension-highlight"
-import Image from "@tiptap/extension-image"
-import Link from "@tiptap/extension-link"
-import Placeholder from "@tiptap/extension-placeholder"
-import Table from "@tiptap/extension-table"
-import TableCell from "@tiptap/extension-table-cell"
-import TableHeader from "@tiptap/extension-table-header"
-import TableRow from "@tiptap/extension-table-row"
-import TaskItem from "@tiptap/extension-task-item"
-import TaskList from "@tiptap/extension-task-list"
-import { useEditor } from "@tiptap/react"
-import StarterKit from "@tiptap/starter-kit"
 import axios from "axios"
 import DOMPurify from "dompurify"
 import _ from "lodash"
 import { marked } from "marked"
 import { useEffect, useState } from "react"
 import { useParams } from "react-router-dom"
-import { Markdown } from "tiptap-markdown"
 
 import MarkdownEditor from "components/pages/MarkdownEditor"
 import PagePreview from "components/pages/PagePreview"
 
-import { EditorContextProvider } from "contexts/EditorContext"
-import { EditorModalContextProvider } from "contexts/EditorModalContext"
+import { useEditorContext } from "contexts/EditorContext"
 
 import { useGetMultipleMediaHook } from "hooks/mediaHooks"
 import { useGetPageHook } from "hooks/pageHooks"
 import { useCspHook } from "hooks/settingsHooks"
 
 import { Editor } from "layouts/components/Editor"
-import Iframe from "layouts/components/Editor/Iframe"
 
 // Isomer utils
 import checkCSP from "utils/cspUtils"
@@ -115,8 +99,9 @@ export const MarkdownEditPage = ({ togglePreview }: MarkdownPageProps) => {
     siteName,
     mediaSrcs,
   })
+  const { editor } = useEditorContext()
 
-  const previewModalProps = useDisclosure()
+  const { onOpen, onClose, ...rest } = useDisclosure()
 
   const { data: csp } = useCspHook()
 
@@ -187,7 +172,11 @@ export const MarkdownEditPage = ({ togglePreview }: MarkdownPageProps) => {
           <Button
             variant="outline"
             alignSelf="center"
-            onClick={previewModalProps.onOpen}
+            onClick={() => {
+              // NOTE: Set initial content to the markdown content first
+              editor.commands.setContent(editorValue)
+              onOpen()
+            }}
           >
             Preview page
           </Button>
@@ -207,8 +196,12 @@ export const MarkdownEditPage = ({ togglePreview }: MarkdownPageProps) => {
       />
 
       <PreviewModal
-        {...previewModalProps}
-        content={editorValue}
+        onClose={() => {
+          onClose()
+          // NOTE: Reset back to what was previously fetched
+          editor.commands.setContent(initialPageData?.content?.pageBody.trim())
+        }}
+        {...rest}
         togglePreview={togglePreview}
       />
     </EditPageLayout>
@@ -218,74 +211,37 @@ export const MarkdownEditPage = ({ togglePreview }: MarkdownPageProps) => {
 interface PreviewModalProps {
   isOpen: boolean
   onClose: () => void
-  content: string
   togglePreview: () => void
 }
 const PreviewModal = ({
   isOpen,
   onClose,
-  content,
   togglePreview,
 }: PreviewModalProps) => {
-  const editor = useEditor({
-    extensions: [
-      StarterKit,
-      Highlight,
-      TaskList,
-      TaskItem,
-      CharacterCount,
-      Image.configure({ allowBase64: true }),
-      Link.configure({ openOnClick: false }),
-      Iframe,
-      Markdown,
-      BubbleMenu.configure({
-        pluginKey: "linkBubble",
-      }),
-      BubbleMenu.configure({
-        pluginKey: "tableBubble",
-      }),
-      Table.configure({
-        resizable: true,
-      }),
-      TableRow,
-      TableHeader,
-      TableCell,
-      Placeholder,
-    ],
-    autofocus: "start",
-    content,
-  })
-
-  useEffect(() => {
-    editor?.commands.setContent(content)
-  }, [content, editor?.commands])
-
-  if (!editor) return null
-
   return (
-    <EditorContextProvider editor={editor}>
-      <EditorModalContextProvider showModal={console.log}>
-        <Modal isOpen={isOpen} onClose={onClose} size="xl">
-          <ModalOverlay />
-          <ModalContent p="1rem">
-            <ModalHeader>
-              <Text textStyle="h3">New editing preview</Text>
-            </ModalHeader>
-            <ModalCloseButton />
-            <ModalBody>
-              <Editor maxW="100%" p="0" />
-            </ModalBody>
-            <ModalFooter>
-              <Button colorScheme="red" mr={3} onClick={onClose}>
-                Close
-              </Button>
-              <Button colorScheme="blue" onClick={togglePreview}>
-                I&apos;m in!
-              </Button>
-            </ModalFooter>
-          </ModalContent>
-        </Modal>
-      </EditorModalContextProvider>
-    </EditorContextProvider>
+    <Modal isOpen={isOpen} onClose={onClose} size="xl">
+      <ModalOverlay />
+      <ModalContent p="1rem">
+        <ModalHeader>
+          <Text textStyle="h3">New editing preview</Text>
+        </ModalHeader>
+        <ModalCloseButton />
+        <ModalBody>
+          <Editor maxW="100%" p="0" />
+          <Alert variant="solid">
+            Changes you make here will reflect on the new editor if you use the
+            new one!
+          </Alert>
+        </ModalBody>
+        <ModalFooter>
+          <Button colorScheme="red" mr={3} onClick={onClose}>
+            Close
+          </Button>
+          <Button colorScheme="blue" onClick={togglePreview}>
+            I&apos;m in!
+          </Button>
+        </ModalFooter>
+      </ModalContent>
+    </Modal>
   )
 }
